+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/settings/censorship/CensorshipViewModel.kt
@@ -3,11 +3,13 @@ package net.nymtech.nymvpn.ui.screens.settings.censorship
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import net.nymtech.nymvpn.data.SettingsRepository
+import net.nymtech.nymvpn.di.qualifiers.ApplicationScope
 import net.nymtech.nymvpn.manager.backend.BackendManager
 import net.nymtech.nymvpn.manager.environment.EnvironmentManager
 import net.nymtech.vpn.backend.Tunnel
@@ -18,6 +20,7 @@ class CensorshipViewModel @Inject constructor(
 	private val backendManager: BackendManager,
 	private val settingsRepository: SettingsRepository,
 	private val environmentManager: EnvironmentManager,
+	@ApplicationScope private val appScope: CoroutineScope,
 ) : ViewModel() {
 
 	private val _uiState = MutableStateFlow(CensorshipUiState())
@@ -37,12 +40,9 @@ class CensorshipViewModel @Inject constructor(
 		_uiState.update { it.copy() }
 	}
 
-	fun reconnect() {
-		viewModelScope.launch {
-			if (backendManager.getState() == Tunnel.State.Up) {
-				backendManager.stopTunnel()
-				backendManager.startTunnel()
-			}
+	fun requestReconnect() {
+		appScope.launch {
+			backendManager.restartTunnel()
 		}
 	}
 }